corrected e2e tests, some improvements
test-e2e
